Move bit_helper to qr2 as only it uses it
