+++ qr2/bit_helper.go
@@ -1,4 +1,4 @@
-package common
+package qr2
 
 import (
 	"errors"
@@ -8,16 +8,16 @@ func aidSlot(aid uint8) uint32 {
 	return 1 << aid
 }
 
-func SetAid(aids uint32, aid uint8) uint32 {
+func setAid(aids uint32, aid uint8) uint32 {
 	return aids | aidSlot(aid)
 }
 
-func ClearAid(aids uint32, aid uint8) uint32 {
+func clearAid(aids uint32, aid uint8) uint32 {
 	return aids & ^aidSlot(aid)
 }
 
 // get the next available aid. pass in the aidBitmap and loop over (at most) 12 aid slot bits, and find the lsb thats 0
-func GetAvailableAid(aidBitmap uint32) (uint8, error) { // TODO: Change to error type or something
+func getAvailableAid(aidBitmap uint32) (uint8, error) { // TODO: Change to error type or something
 	var i uint8
 	for i = range 12 {
 		if ((aidBitmap >> i) & 1) == 0 {
@@ -29,10 +29,10 @@ func GetAvailableAid(aidBitmap uint32) (uint8, error) { // TODO: Change to error
 	return 0xff, errors.New("No available aid!")
 }
 
-func SetLocalPlayerCount(localPlayerCount uint32) uint32 {
+func setLocalPlayerCount(localPlayerCount uint32) uint32 {
 	return localPlayerCount << 24
 }
 
-func GetAidPlayerCount(playerCount uint32) uint32 {
+func getAidPlayerCount(playerCount uint32) uint32 {
 	return playerCount >> 24
 }
